# docs/api/openapi.yaml
openapi: 3.0.0
info:
  title: Label Computation System API
  version: 1.0.0
  description: |
    High-performance label computation system for quantitative trading that processes support/resistance level events 
    and computes forward-looking labels for pattern mining and ML models.
    
    Key features:
    - Dual-mode processing: Batch backfill (minutes) and real-time incremental (<1 second)
    - Multi-timeframe alignment using lower granularity data for path-dependent calculations
    - Support for 29 FX pairs and indices with billions of labels
    
    Critical SLAs:
    - Incremental compute: <100ms p99 latency
    - Batch throughput: 1M+ candles/minute
    - Cache hit rate: >95% for recent 24 hours
  contact:
    name: Label Computation System Team
    email: support@labelcompute.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.labelcompute.com/v1
    description: Production server
  - url: https://staging-api.labelcompute.com/v1
    description: Staging server
  - url: http://localhost:8000/v1
    description: Development server
security:
  - bearerAuth: []
  - apiKey: []
...
paths:
...
  /cache/stats:
    get:
      tags: [Cache]
      summary: Cache statistics
      description: Get detailed cache performance statistics
      operationId: getCacheStats
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Cache stats retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CacheStatsResponse'
  /cache/warm:
    post:
      tags: [Cache]
      summary: Warm cache for instrument
      description: Pre-load cache with recent data for an instrument
      operationId: warmCache
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                instrument_id:
                  type: string
                  description: Instrument to warm cache for
                granularity:
                  type: string
                  enum: [M15, H1, H4, D, W]
                hours:
                  type: integer
                  minimum: 1
                  maximum: 168
                  default: 24
                  description: Hours of data to cache
              required: [instrument_id, granularity]
            example:
              instrument_id: "EURUSD"
              granularity: "H4"
              hours: 24
      responses:
        '202':
          description: Cache warming started
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Cache warming started for EURUSD H4"
                  estimated_completion:
                    type: string
                    format: date-time
  /cache/invalidate:
    delete:
      tags: [Cache]
      summary: Invalidate cache entries
      description: Clear cache for specific instrument or global patterns
      operationId: invalidateCache
      security:
        - bearerAuth: []
      parameters:
        - name: instrument_id
          in: query
          description: Instrument to invalidate (optional)
          required: false
          schema:
            type: string
        - name: granularity
          in: query
          description: Granularity to invalidate (optional)
          required: false
          schema:
            type: string
            enum: [M15, H1, H4, D, W]
        - name: pattern
          in: query
          description: Redis key pattern to match
          required: false
          schema:
            type: string
          example: "labels:EURUSD:H4:*"
      responses:
        '200':
          description: Cache invalidated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Invalidated 1250 cache entries"
                  keys_deleted:
                    type: integer
                    example: 1250

# ===========================================================================
# COMPONENTS
# ===========================================================================
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token containing user claims and permissions.
        Header format: `Authorization: Bearer <token>`
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        API key for service-to-service authentication.
        Header format: `X-API-Key: <key>`
  schemas:
...
    CacheStatsResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        redis:
          type: object
          properties:
            memory_usage_mb:
              type: number
              format: double
            keys_total:
              type: integer
            hit_rate:
              type: number
              format: double
              minimum: 0
              maximum: 1
            evictions_last_hour:
              type: integer
            connections_active:
              type: integer
        cache_levels:
          type: object
          properties:
            labels:
              type: object
              properties:
                keys_count:
                  type: integer
                hit_rate:
                  type: number
                  format: double
                avg_ttl_minutes:
                  type: number
                  format: double
            lookback_data:
              type: object
              properties:
                keys_count:
                  type: integer
                hit_rate:
                  type: number
                  format: double
                avg_ttl_minutes:
                  type: number
                  format: double
            levels:
              type: object
              properties:
                keys_count:
                  type: integer
                hit_rate:
                  type: number
                  format: double
                avg_ttl_minutes:
                  type: number
                  format: double
...