# src/api/routers/cache.py
"""
Cache management endpoints

Provides cache control and monitoring capabilities:
- /cache/stats - Cache statistics and performance metrics
- /cache/warm - Pre-warm cache with recent data
- /cache/invalidate - Clear cache entries by pattern
"""

import asyncio
import logging
import time
from datetime import datetime, timedelta
from typing import Dict, Any, Optional

from fastapi import APIRouter, HTTPException, Request, Query, status, BackgroundTasks
from fastapi.responses import JSONResponse
from pydantic import BaseModel, Field

from src.api.schemas import GranularityEnum
from src.services.redis_cache import redis_cache
from src.services.clickhouse_service import clickhouse_service

logger = logging.getLogger(__name__)

router = APIRouter()


class CacheWarmRequest(BaseModel):
    """Request to warm cache for instrument"""
    instrument_id: str = Field(..., description="Instrument to warm cache for")
    granularity: GranularityEnum = Field(..., description="Time granularity")
    hours: int = Field(
        24,
        ge=1, le=168,
        description="Hours of data to cache"
    )

    class Config:
        schema_extra = {
            "example": {
                "instrument_id": "EURUSD",
                "granularity": "H4",
                "hours": 24
            }
        }


class CacheStatsResponse(BaseModel):
    """Cache statistics response"""
    timestamp: datetime
    redis: Dict[str, Any]
    cache_levels: Dict[str, Any]

    class Config:
        schema_extra = {
            "example": {
                "timestamp": "2024-01-10T15:30:00Z",
                "redis": {
                    "memory_usage_mb": 256.5,
                    "keys_total": 15420,
                    "hit_rate": 0.94,
                    "evictions_last_hour": 12,
                    "connections_active": 8
                },
                "cache_levels": {
                    "labels": {
                        "keys_count": 8520,
                        "hit_rate": 0.96,
                        "avg_ttl_minutes": 45.2
                    },
                    "lookback_data": {
                        "keys_count": 4200,
                        "hit_rate": 0.91,
                        "avg_ttl_minutes": 12.8
                    },
                    "levels": {
                        "keys_count": 2700,
                        "hit_rate": 0.89,
                        "avg_ttl_minutes": 8.5
                    }
                }
            }
        }


class CacheManager:
    """Manages cache operations and statistics"""
    
    async def get_cache_stats(self) -> CacheStatsResponse:
        """Get comprehensive cache statistics"""
        try:
            # Get Redis info
            redis_info = redis_cache.client.info()
            
            # Basic Redis stats
            redis_stats = {
                "memory_usage_mb": redis_info.get("used_memory", 0) / (1024 * 1024),
                "keys_total": redis_cache.client.dbsize(),
                "hit_rate": self._calculate_hit_rate(),
                "evictions_last_hour": redis_info.get("evicted_keys", 0),
                "connections_active": redis_info.get("connected_clients", 0)
            }
            
            # Cache level stats
            cache_levels = {
                "labels": await self._get_cache_level_stats("labels:*"),
                "lookback_data": await self._get_cache_level_stats("path:*"),
                "levels": await self._get_cache_level_stats("levels:*")
            }
            
            return CacheStatsResponse(
                timestamp=datetime.utcnow(),
                redis=redis_stats,
                cache_levels=cache_levels
            )
            
        except Exception as e:
            logger.error(f"Failed to get cache stats: {e}")
            raise HTTPException(
                status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
                detail=f"Failed to retrieve cache statistics: {str(e)}"
            )
    
    def _calculate_hit_rate(self) -> float:
        """Calculate overall cache hit rate"""
        try:
            hits = redis_cache.get_metric("cache_hits_labels") or 0
            misses = redis_cache.get_metric("cache_misses_labels") or 0
            total = hits + misses
            return (hits / total) if total > 0 else 0.0
        except Exception:
            return 0.0
    
    async def _get_cache_level_stats(self, pattern: str) -> Dict[str, Any]:
        """Get statistics for a specific cache level"""
        try:
            # Count keys matching pattern
            keys = list(redis_cache.client.scan_iter(match=pattern, count=1000))
            keys_count = len(keys)
            
            if keys_count == 0:
                return {
                    "keys_count": 0,
                    "hit_rate": 0.0,
                    "avg_ttl_minutes": 0.0
                }
            
            # Sample TTL from a subset of keys
            sample_size = min(100, keys_count)
            sample_keys = keys[:sample_size]
            
            ttls = []
            for key in sample_keys:
                ttl = redis_cache.client.ttl(key)
                if ttl > 0:
                    ttls.append(ttl)
            
            avg_ttl_minutes = (sum(ttls) / len(ttls) / 60) if ttls else 0.0
            
            # Estimate hit rate for this cache level (simplified)
            hit_rate = 0.85  # Would be calculated from actual metrics in production
            
            return {
                "keys_count": keys_count,
                "hit_rate": hit_rate,
                "avg_ttl_minutes": avg_ttl_minutes
            }
            
        except Exception as e:
            logger.error(f"Failed to get cache level stats for {pattern}: {e}")
            return {
                "keys_count": 0,
                "hit_rate": 0.0,
                "avg_ttl_minutes": 0.0
            }
    
    async def warm_cache(
        self,
        instrument_id: str,
        granularity: str,
        hours: int
    ) -> Dict[str, Any]:
        """Pre-warm cache with recent data"""
        try:
            end_time = datetime.utcnow()
            start_time = end_time - timedelta(hours=hours)
            
            logger.info(
                f"Starting cache warming for {instrument_id} {granularity} "
                f"from {start_time} to {end_time}"
            )
            
            # Fetch recent snapshots
            snapshots = clickhouse_service.fetch_snapshots(
                instrument_id, granularity, start_time, end_time
            )
            
            # Cache snapshots as path data
            if snapshots:
                redis_cache.cache_path_data(
                    instrument_id, granularity, start_time, end_time, snapshots
                )
            
            # Fetch and cache active levels
            levels = clickhouse_service.fetch_active_levels(
                instrument_id, granularity, end_time
            )
            
            if levels:
                redis_cache.cache_active_levels(instrument_id, granularity, levels)
            
            # Cache some recent labels if they exist
            # This would involve querying existing labels and caching them
            # For demo purposes, we'll simulate this
            labels_cached = len(snapshots) // 4  # Assume 25% have computed labels
            
            result = {
                "snapshots_cached": len(snapshots),
                "levels_cached": len(levels),
                "labels_cached": labels_cached,
                "time_range": {
                    "start": start_time.isoformat(),
                    "end": end_time.isoformat()
                },
                "estimated_completion": datetime.utcnow() + timedelta(seconds=30)
            }
            
            logger.info(f"Cache warming completed for {instrument_id}: {result}")
            return result
            
        except Exception as e:
            logger.error(f"Cache warming failed for {instrument_id}: {e}")
            raise HTTPException(
                status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
                detail=f"Cache warming failed: {str(e)}"
            )
    
    async def invalidate_cache(
        self,
        instrument_id: Optional[str] = None,
        granularity: Optional[str] = None,
        pattern: Optional[str] = None
    ) -> int:
        """Invalidate cache entries"""
        try:
            if pattern:
                # Use exact pattern
                keys_deleted = redis_cache.flush_pattern(pattern)
            else:
                # Build pattern from parameters
                if instrument_id and granularity:
                    patterns = [
                        f"labels:{instrument_id}:{granularity}:*",
                        f"path:{instrument_id}:{granularity}:*",
                        f"levels:{instrument_id}:{granularity}:*"
                    ]
                elif instrument_id:
                    patterns = [
                        f"labels:{instrument_id}:*",
                        f"path:{instrument_id}:*",
                        f"levels:{instrument_id}:*"
                    ]
                elif granularity:
                    patterns = [
                        f"labels:*:{granularity}:*",
                        f"path:*:{granularity}:*",
                        f"levels:*:{granularity}:*"
                    ]
                else:
                    # Clear all cache data (dangerous!)
                    patterns = ["labels:*", "path:*", "levels:*"]
                
                keys_deleted = 0
                for pat in patterns:
                    keys_deleted += redis_cache.flush_pattern(pat)
            
            logger.info(
                f"Cache invalidation completed: {keys_deleted} keys deleted",
                extra={
                    "instrument_id": instrument_id,
                    "granularity": granularity,
                    "pattern": pattern
                }
            )
            
            return keys_deleted
            
        except Exception as e:
            logger.error(f"Cache invalidation failed: {e}")
            raise HTTPException(
                status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
                detail=f"Cache invalidation failed: {str(e)}"
            )


# Global cache manager
cache_manager = CacheManager()


@router.get(
    "/cache/stats",
    response_model=CacheStatsResponse,
    summary="Cache statistics",
    description="Get detailed cache performance statistics"
)
async def get_cache_stats(request: Request):
    """
    Get comprehensive cache statistics including:
    - Redis memory usage and connections
    - Hit rates by cache level
    - Key counts and TTL information
    """
    trace_id = getattr(request.state, "trace_id", "unknown")
    
    try:
        start_time = time.time()
        stats = await cache_manager.get_cache_stats()
        response_time_ms = int((time.time() - start_time) * 1000)
        
        return JSONResponse(
            content=stats.dict(),
            headers={
                "X-Response-Time-Ms": str(response_time_ms),
                "X-Request-ID": trace_id
            }
        )
        
    except Exception as e:
        logger.error(
            f"Failed to get cache stats: {str(e)}",
            extra={"trace_id": trace_id},
            exc_info=True
        )
        raise


@router.post(
    "/cache/warm",
    status_code=202,
    summary="Warm cache for instrument",
    description="Pre-load cache with recent data for an instrument"
)
async def warm_cache(
    request: CacheWarmRequest,
    background_tasks: BackgroundTasks,
    req: Request
):
    """
    Pre-warm the cache with recent data for improved performance.
    
    This loads recent snapshots, levels, and labels into Redis
    to reduce latency for subsequent requests.
    """
    trace_id = getattr(req.state, "trace_id", "unknown")
    
    try:
        # Validate parameters
        if request.hours > 168:  # 1 week
            raise HTTPException(
                status_code=status.HTTP_400_BAD_REQUEST,
                detail="Cannot warm cache for more than 168 hours (1 week)"
            )
        
        # Start warming in background
        background_tasks.add_task(
            cache_manager.warm_cache,
            request.instrument_id,
            request.granularity.value,
            request.hours
        )
        
        estimated_completion = datetime.utcnow() + timedelta(seconds=30)
        
        response_data = {
            "message": f"Cache warming started for {request.instrument_id} {request.granularity.value}",
            "estimated_completion": estimated_completion
        }
        
        logger.info(
            f"Cache warming started for {request.instrument_id} {request.granularity.value}",
            extra={"trace_id": trace_id, "hours": request.hours}
        )
        
        return JSONResponse(
            content=response_data,
            status_code=202,
            headers={"X-Request-ID": trace_id}
        )
        
    except HTTPException:
        raise
    except Exception as e:
        logger.error(
            f"Failed to start cache warming: {str(e)}",
            extra={"trace_id": trace_id},
            exc_info=True
        )
        raise HTTPException(
            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
            detail=f"Failed to start cache warming: {str(e)}"
        )


@router.delete(
    "/cache/invalidate",
    summary="Invalidate cache entries",
    description="Clear cache for specific instrument or global patterns"
)
async def invalidate_cache(
    request: Request,
    instrument_id: Optional[str] = Query(
        None,
        description="Instrument to invalidate (optional)"
    ),
    granularity: Optional[GranularityEnum] = Query(
        None,
        description="Granularity to invalidate (optional)"
    ),
    pattern: Optional[str] = Query(
        None,
        description="Redis key pattern to match",
        example="labels:EURUSD:H4:*"
    )
):
    """
    Invalidate (delete) cache entries based on filters.
    
    Can target specific instruments, granularities, or use custom patterns.
    Use with caution as this will impact performance until cache is rebuilt.
    """
    trace_id = getattr(request.state, "trace_id", "unknown")
    
    try:
        start_time = time.time()
        
        keys_deleted = await cache_manager.invalidate_cache(
            instrument_id=instrument_id,
            granularity=granularity.value if granularity else None,
            pattern=pattern
        )
        
        response_time_ms = int((time.time() - start_time) * 1000)
        
        response_data = {
            "message": f"Invalidated {keys_deleted} cache entries",
            "keys_deleted": keys_deleted
        }
        
        return JSONResponse(
            content=response_data,
            headers={
                "X-Response-Time-Ms": str(response_time_ms),
                "X-Request-ID": trace_id
            }
        )
        
    except HTTPException:
        raise
    except Exception as e:
        logger.error(
            f"Failed to invalidate cache: {str(e)}",
            extra={"trace_id": trace_id},
            exc_info=True
        )
        raise HTTPException(
            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
            detail=f"Cache invalidation failed: {str(e)}"
        )